% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.additivehierbasis.R
\name{cv.additivehierbasis}
\alias{cv.additivehierbasis}
\title{Cross-validation for Multivariate additivehierbasis Models}
\usage{
cv.additivehierbasis(X, y, lambdas = NULL, nfolds = 10, ...)
}
\arguments{
\item{X}{An \eqn{n \times p}{n x p} matrix of covariates.}

\item{y}{A univariate vector respresenting the response.}

\item{nfolds}{Number of cross-validation folds. Default: \code{nfolds = 10}.}

\item{...}{Other arguments that may be passed to \code{hierbasis}.}

\item{lambda}{(Optional) User-specified sequence of tuning parameters \eqn{lambda}.}
}
\value{
Returns an object of class \code{hierbasis} with elements (to finish...)

\item{to do...}{to do...}
}
\description{
Cross-validation for Multivariate additivehierbasis Models
}
\details{
The function runs \code{nfolds} + 1 times. If \code{lambdas}
         is not specified then it uses the first run to find the
         sequence of \code{lambdas}. The remaining \code{nfolds}
         runs are done using the fold subsets as usual.
}
\references{
Haris, A., Shojaie, A. and Simon, N. (2016). Nonparametric Regression with
Adaptive Smoothness via a Convex Hierarchical Penalty. Available on request
by authors.
}
\author{
Annik Gougeon,
David Fleischer (\email{david.fleischer@mail.mcgill.ca}).
}
