% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierbasis.R
\name{hierbasis}
\alias{hierbasis}
\title{Univariate Nonparametric Regression Estimation
via a Hierarchical Penalty}
\usage{
hierbasis(x, y, nbasis = length(y), max.lambda = NULL,
  lam.min.ratio = 1e-04, nlam = 50, m = 3, type = c("gaussian",
  "binomial"))
}
\arguments{
\item{x}{A univariate vector representing the predictor.}

\item{y}{A univariate vector respresenting the response.}

\item{nbasis}{The number of basis functions to be used in the
basis expansion of \code{x}. Default: \code{nbasis = length(y)}.}

\item{max.lambda}{The largest value of \eqn{\lambda} penalizing
the hierarchical penalty. Default: \code{max.lambda = NULL}
(computed data-adaptively).}

\item{lam.min.ratio}{The ratio of the smallest value of \eqn{\lambda}
to the largest \eqn{\lambda}. Default: \code{lam.min.ratio = 1e-04}.}

\item{nlam}{The number \eqn{\lambda}'s to compute the \code{hierbasis}
estimators (computed on a base-10 \eqn{log} scale). Default:
\code{nlam = 50}.}

\item{m}{Smoothing parameter controlling the degree of polynomial
smoothing/decay performed by the weights in the penalty \eqn{\Omega}.
Default: \code{m = 3}.}

\item{type}{Specifies either Gaussian regression (\code{"gaussian"})
with a continuous response vector or binomial regression
(\code{"binomial"}) with binary response. Default type is assumed to be
Gaussian.}
}
\value{
Returns an object of class \code{hierbasis} with elements
}
\description{
Univariate Nonparametric Regression Estimation
via a Hierarchical Penalty
}
\details{
Solves the univariate minimization problem (see Haris et al. 2016)
\deqn{argmin_{\beta} (1/2n) || y - \Psi_K \beta ||^2_2 + \lambda\Omega(\beta; m),}
where \eqn{\beta} is a real-valued vector of length \eqn{K = } \code{nbasis},
\eqn{\Psi_K} is a \eqn{n\times K} real-valued matrix whose columns
correspond to the basis expansion of \eqn{x} (ordered in increasing)
complexity. The penalty function \eqn{\Omega} is defined as
\deqn{\Omega(\beta; m) = \sum^K_{k = 1} w_{k, m} || \Psi_{k:K} \beta_{k:K} ||_2,}
for \eqn{\Psi_{k:K}} denotes the submatrix of \eqn{\Psi_K} corresponding to
the final columns \eqn{k, k + 1, ..., K}, and \eqn{\beta_{k:K}}
the subvector of \eqn{\beta} corresponding to the final elements
\eqn{k, k + 1, ..., K}. The weights \eqn{w_{k, m}} in \eqn{\Omega}
are given by the m - 1 degree polynomial
\deqn{w_{k, m} = k^m - (k - 1)^m},
controlling the level of decay on higher order estimates of \eqn{\beta}.
}
\references{
Haris, A., Shojaie, A. and Simon, N. (2016). Nonparametric Regression with
Adaptive Smoothness via a Convex Hierarchical Penalty. Available on request
by authors.
}
\seealso{
The original \code{HierBasis} function, as implemented by
Haris et al. (2016) can be found via
\url{https://github.com/asadharis/HierBasis/}.
}
\author{
Annik Gougeon,
David Fleischer (\email{david.fleischer@mail.mcgill.ca}).
}
