% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.additivehierbasis.R
\name{predict.additivehierbasis}
\alias{predict.additivehierbasis}
\title{Model Predictions for the Multivariate additivehierbasis model}
\usage{
\method{predict}{additivehierbasis}(object, new.X = NULL, ...)
}
\arguments{
\item{object}{A fitted object of class '\code{additivehierbasis}'.}

\item{...}{Not used. Other arguments for predict function.}

\item{new.x}{An optional matrix of values of \code{x} at which to predict.
The number of columns of \code{new.x} should be equal to the number of
columns of \code{object$x}.}
}
\value{
\item{fitted.values}{A matrix of fitted values with \code{nrow(new.x)}
                     rows and \code{nlam} columns}
}
\description{
The generic S3 method for predictions for objects of
class \code{additivehierbasis}.
}
\details{
This function returns a matrix of  predicted values at the specified
values of x given by \code{new.x}. Each column corresponds to a lambda value
used for fitting the original model.

If \code{new.x == NULL} then this function simply returns
the fitted values of the estimated function at the original x values used for
model fitting. The predicted values are presented for each lambda values.
}
\references{
Haris, A., Shojaie, A. and Simon, N. (2016). Nonparametric Regression with
Adaptive Smoothness via a Convex Hierarchical Penalty. Available on request
by authors.
}
\seealso{
The original \code{HierBasis} function, as implemented by
Haris et al. (2016) can be found via
\url{https://github.com/asadharis/HierBasis/}.
}
\author{
Annik Gougeon,
David Fleischer (\email{david.fleischer@mail.mcgill.ca}).
}
